Siddharth 03

+++ app/src/main/java/com/e/hardwarewalaseller/beans/ItemList.java
@@ -42,7 +42,7 @@ public class ItemList implements Serializable {
     private String shopkeeperId;
     @SerializedName("total")
     @Expose
-    private Long total;
+    private Double total;
     @SerializedName("userId")
     @Expose
     private String userId;
@@ -131,11 +131,11 @@ public class ItemList implements Serializable {
 
     public void setshopkeeperId(String shopkeeperId) { this.shopkeeperId = shopkeeperId; }
 
-    public Long getTotal() {
+    public Double getTotal() {
         return total;
     }
 
-    public void setTotal(Long total) {
+    public void setTotal(Double total) {
         this.total = total;
     }
